requirements.txt
libclang
argparse
pytest
rich
loguru
setuptools
wheel
pyinstaller